# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maktab Raqamli Tizimi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - in: query
          name: studentId
          schema:
            type: number
        - in: query
          name: classId
          schema:
            type: number
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Mark attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      parameters:
        - in: query
          name: classId
          schema:
            type: number
        - in: query
          name: subjectId
          schema:
            type: number
        - in: query
          name: teacherId
          schema:
            type: number
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get assignment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update assignment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete assignment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - in: query
          name: classId
          schema:
            type: number
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get announcement
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, firstName, lastName, role, username, createdAt]
      properties:
        id:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [director, zavuch, teacher, student, parent]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        parentId:
          type: ["integer", "null"]
        subjectIds:
          type: array
          items:
            type: number
        createdAt:
          type: string
          format: date-time
...
    Attendance:
      type: object
      required: [id, studentId, classId, date, status, createdAt]
      properties:
        id:
          type: number
        studentId:
          type: number
        studentName:
          type: string
        classId:
          type: number
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late, excused]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AttendanceInput:
      type: object
      required: [studentId, classId, date, status]
      properties:
        studentId:
          type: number
        classId:
          type: number
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late, excused]
        note:
          type: string
    AttendanceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [present, absent, late, excused]
        note:
          type: string
    Assignment:
      type: object
      required: [id, title, subjectId, teacherId, classId, dueDate, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: ["string", "null"]
        subjectId:
          type: number
        subjectName:
          type: string
        teacherId:
          type: number
        classId:
          type: number
        dueDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    AssignmentInput:
      type: object
      required: [title, subjectId, teacherId, classId, dueDate]
      properties:
        title:
          type: string
        description:
          type: string
        subjectId:
          type: number
        teacherId:
          type: number
        classId:
          type: number
        dueDate:
          type: string
          format: date
    AssignmentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date
    Announcement:
      type: object
      required: [id, title, content, authorId, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        content:
          type: string
        authorId:
          type: number
        authorName:
          type: string
        classId:
          type: ["integer", "null"]
        priority:
          type: string
          enum: [normal, high, urgent]
        createdAt:
          type: string
          format: date-time
    AnnouncementInput:
      type: object
      required: [title, content, authorId]
      properties:
        title:
          type: string
        content:
          type: string
        authorId:
          type: number
        classId:
          type: number
        priority:
          type: string
          enum: [normal, high, urgent]
          default: normal
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        priority:
          type: string
          enum: [normal, high, urgent]
...