# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Maktab Raqamli Tizimi API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementUpdatePriority } from './announcementUpdatePriority';

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  priority?: AnnouncementUpdatePriority;
}